Feature fix status delete

+++ app/handler/statuses/delete.go
@@ -1,7 +1,10 @@
 package statuses
 
 import (
+	"encoding/json"
+	"fmt"
 	"net/http"
+	"yatter-backend-go/app/handler/auth"
 	"yatter-backend-go/app/handler/httperror"
 	"yatter-backend-go/app/handler/request"
 )
@@ -16,7 +19,31 @@ func (h *handler) Delete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	status, err := h.app.Dao.Status().FindByID(ctx, id)
+	if err != nil {
+		httperror.InternalServerError(w, err)
+		return
+	}
+	if status == nil {
+		httperror.Error(w, http.StatusNotFound)
+		return
+	}
+
+	login := auth.AccountOf(r)
+	if login == nil {
+		httperror.InternalServerError(w, fmt.Errorf("lost account"))
+		return
+	}
+	if status.Account.ID != login.ID {
+		httperror.BadRequest(w, fmt.Errorf("the status does not belong to loginUser"))
+	}
+
 	if err = h.app.Dao.Status().Delete(ctx, id); err != nil {
+		httperror.InternalServerError(w, err)
+		return
+	}
+
+	if err := json.NewEncoder(w).Encode(&struct{}{}); err != nil {
 		httperror.InternalServerError(w, err)
 		return
 	}
